Moved getType
