+++ scripts/public_api.mjs
@@ -1,4 +1,4 @@
-import { MODULE, TYPES } from "./constants.mjs";
+import { MODULE } from "./constants.mjs";
 import { _splitTokensByDisposition } from "./helpers/auraHelpers.mjs";
 import {
   _getMinimumDistanceBetweenTokens,
@@ -6,7 +6,8 @@ import {
   _getAppId,
   _createBabonus,
   _openWorkshop,
-  _getAllTokenGridSpaces
+  _getAllTokenGridSpaces,
+  _getType
 } from "./helpers/helpers.mjs";
 import { _getAllContainingTemplates } from "./helpers/templateHelpers.mjs";
 import { migration } from "./migration.mjs";
@@ -105,17 +106,8 @@ function getBonuses() {
  * Returns the bonuses of a given type.
  * Returned in the form of [id, values].
  */
-export function getType(object, type) {
-  if (!TYPES.map(t => t.value).includes(type)) {
-    console.error(`'${type}' is not a valid Build-a-Bonus type!`);
-    return null;
-  }
-  const flag = object.getFlag(MODULE, "bonuses") ?? {};
-  return Object.entries(flag).filter(([id, values]) => {
-    const validId = foundry.data.validators.isValidId(id);
-    const validtype = values?.type === type;
-    return validId && validtype;
-  });
+function getType(object, type) {
+  return _getType(object, type);
 }
 
 /**
